Title: 재귀함수를 이용한 피보나치 수, Time: 8ms, Memory: 0MB, Status: Passed - Codetree

+++ 251018/재귀함수를 이용한 피보나치 수/fibonacci-using-recursive-function.c
@@ -10,7 +10,7 @@ int main() {
     return 0;
 }
 
-long fibonacci(int n) {
+int fibonacci(int n) {
     if (n < 0) return 0;
     if (n == 1 || n == 2) return 1;
